add html and fix bug

+++ src/main/java/org/jplus/mapper/BksktjxMapper.java
@@ -25,5 +25,5 @@ public interface BksktjxMapper {
 
     void deleteClassInfoBybksktjxId(Integer bksktjx);
 
-    Integer getBksktjxSum(String gh);
+    Float getBksktjxSum(String gh);
 }
